create modal code organisation
